Improve eventBridge entry point API
This change brings the API for eventBridge closer to that seen in the code snippet on https://notation.dev by:

* Renaming 'ruleName' config option to 'name'
* Accepting schedule handler function in the config object passed in as the 'handler' field.
* Renamed the schedule builder functions (to 'rate', 'once', 'cron'.)

Additional changes:
* No longer export 'toAwsScheduleExpression' from the module as this is a notation helper function. It has been moved to its own file.
