+++ packages/aws/src/event-bridge/event-bridge-schedule.ts
@@ -1,28 +1,30 @@
 import { ResourceGroup } from "@notation/core";
 import { EventBridgeHandler } from "src/shared/lambda.handler";
-import { Schedule, toAwsScheduleExpression } from "./schedule";
+import { Schedule } from "./schedule";
 import { lambda } from "src/lambda";
 import * as aws from "@notation/aws.iac";
+import { toAwsScheduleExpression } from "./aws-conversions";
 
-export const eventBridgeSchedule = (
-  config: { ruleName: string; schedule: Schedule },
-  handler: EventBridgeHandler<"Scheduled Event", any>,
-): ResourceGroup => {
+export const schedule = (config: {
+  name: string;
+  schedule: Schedule;
+  handler: EventBridgeHandler<"Scheduled Event", any>;
+}): ResourceGroup => {
   const eventBridgeScheduleGroup = new aws.AwsResourceGroup(
     "aws/eventBridge/schedule",
     config,
   );
 
   // at compile time becomes infra module
-  const lambdaGroup = handler as any as ReturnType<typeof lambda>;
+  const lambdaGroup = config.handler as any as ReturnType<typeof lambda>;
   const lambdaResource = lambdaGroup.findResource(aws.lambda.LambdaFunction)!;
 
   const eventBridgeRule = new aws.eventBridge.EventBridgeRule({
-    id: `${config.ruleName}-eventbridge-rule`,
+    id: `${config.name}-eventbridge-rule`,
     config: {
-      Name: config.ruleName,
+      Name: config.name,
       ScheduleExpression: toAwsScheduleExpression(config.schedule),
-      EventBusName: "default"
+      EventBusName: "default",
     },
     dependencies: {
       lambda: lambdaResource,
@@ -38,7 +40,7 @@ export const eventBridgeSchedule = (
   if (!permission) {
     lambdaGroup.add(
       new aws.eventBridge.LambdaEventBridgeRulePermission({
-        id: `${config.ruleName}-eventbridge-permission`,
+        id: `${config.name}-eventbridge-permission`,
         dependencies: {
           lambda: lambdaResource,
           eventBridgeRule: eventBridgeRule,
