+++ test/compiler.test.app/infra/event-bridge.ts
@@ -1,10 +1,8 @@
-import { eventBridgeSchedule, rateSchedule } from "@notation/aws/event-bridge";
+import * as eventBridge from "@notation/aws/event-bridge";
 import { exampleHandler2 } from "runtime/eventbridge/scheduleEvent.fn";
 
-eventBridgeSchedule(
-  {
-    ruleName: "Event-bridge-schedule-test",
-    schedule: rateSchedule(1, "minute"),
-  },
-  exampleHandler2,
-);
+eventBridge.schedule({
+  name: "Event-bridge-schedule-test",
+  schedule: eventBridge.rate(1, "minute"),
+  handler: exampleHandler2,
+});
